requirements.txt
Flask==2.3.0
numpy<2
tensorflow==2.13.0
opencv-python==4.8.0.74
Pillow==10.0.0
scikit-learn==1.3.0
groq==0.4.1
python-dotenv==1.0.0
werkzeug==2.3.0
